+++ lazy5/ui/QtHdfLoad.py
@@ -80,21 +80,32 @@ class HdfLoad(_QDialog): ### EDIT ###
 
         # pragma: no cover
         dialog = HdfLoad(parent)
-        ret_fileopen = dialog.fileOpen(pth)
 
-        ret = None
-        if ret_fileopen:
-            ret_dset_select = dialog.exec_()
-            if ret_dset_select == _QDialog.Rejected:
-                pass
-            elif dialog.all_selected is None:
-                pass
+        ret_fileopen = True
+        pth = _os.path.abspath(pth)
+
+        while True:
+            ret_fileopen = dialog.fileOpen(pth)
+
+            ret = None
+            if ret_fileopen:
+                ret_dset_select = dialog.exec_()
+                if ret_dset_select == _QDialog.Rejected:
+                    pth = dialog.path
+                elif dialog.all_selected is None:
+                    pass
+                else:
+                    ret = (dialog.path, dialog.filename, dialog.all_selected)
+                    break
             else:
-                ret = (dialog.path, dialog.filename, dialog.all_selected)
+                break
         return ret
 
     def fileOpen(self, pth='./'):  # Qt-related pylint: disable=C0103
         """ Select HDF5 File via QDialog built-in."""
+
+        if pth is None:
+            pth = './'
 
         if _os.path.isdir(pth):  # No file provided, use QFileDialog; # pragma: no cover
             filetype_options = 'HDF5 Files (*.h5 *.hdf);;All Files (*.*)'
@@ -129,7 +140,9 @@ class HdfLoad(_QDialog): ### EDIT ###
 
         #self.dsetlist = QListWidget(self.verticalLayoutWidget)
         self.ui.listDataSet.clear()
-        self.ui.listDataSet.addItems(self.group_dset_dict[self.ui.comboBoxGroupSelect.currentText()])
+
+        if self.ui.comboBoxGroupSelect.currentText() != '':
+            self.ui.listDataSet.addItems(self.group_dset_dict[self.ui.comboBoxGroupSelect.currentText()])
         #print('Changed')
 
     def populate_attrs(self, attr_dict=None):