+++ lazy5/utils.py
@@ -1,7 +1,13 @@
 """ Utility functions """
-import h5py as _h5py
+import os as _os
 
-__all__ = ['FidOrFile', 'hdf_is_open']
+import h5py as _h5py
+import numpy as _np
+
+from .config import DefaultConfig
+_h5py.get_config().complex_names = DefaultConfig().complex_names
+
+__all__ = ['FidOrFile', 'hdf_is_open', 'fullpath']
 
 class FidOrFile:
     """
@@ -73,3 +79,10 @@ def hdf_is_open(fid):
         return True
     else:
         return None
+
+def fullpath(filename, pth=None):
+    """ Return a full path by joining a pth and filename """
+    if not pth:
+        return filename
+    else:
+        return _os.path.join(pth, filename)