+++ CHANGELOG.rst
@@ -2,6 +2,16 @@
 Changelog
 =========
 
+0.2.0 (18-07-09)
+--------
+
+- Alter attributes with and without ensuring same type
+- Save dataset to file with attributes and chunking
+- Write attributes via dictionary
+- Macros to check file and dataset validity
+- Improved handling of strings in numpy array in attributes
+- Can convert single-entry numpy array numeric attributes to Python native types
+
 0.1.0 (18-05-04)
 ----------------